fix(lineage): Move lineage registry to entity-registry module (#4412)
* Move lineage registry

* Fix test
